requirements.txt
absl-py==2.1.0
accelerate==0.28.0
bleach==1.5.0
blenderproc==2.7.0
cachetools==5.3.3
contourpy==1.1.1
cycler==0.12.1
diffusers==0.14.0
easydict==1.13
filelock==3.13.1
fonttools==4.49.0
freetype-py==2.4.0
fsspec==2024.2.0
google-auth==2.28.2
google-auth-oauthlib==1.0.0
grpcio==1.62.1
h5py==3.10.0
html5lib==0.9999999
huggingface-hub==0.21.4
imageio==2.34.0
importlib-metadata==7.0.2
importlib-resources==6.3.0
kiwisolver==1.4.5
markdown==3.3.4
markupsafe==2.1.5
matplotlib==3.7.5
networkx==3.1
numpy==1.20.3
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==24.0
pandas==2.0.3
protobuf==5.26.0
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyglet==2.0.14
pyopengl==3.1.0
pyparsing==3.1.2
pyrender==0.1.45
python-dateutil==2.9.0.post0
regex==2023.12.25
requests-oauthlib==1.4.0
rsa==4.9
ruamel-yaml==0.17.16
ruamel-yaml-clib==0.2.6
safetensors==0.4.2
scipy==1.10.1
tqdm==4.66.2
transformers==4.27.4
trimesh==4.2.0
tzdata==2024.1
pillow==10.2.0
progressbar==2.5
pyyaml==5.4.1
zipp==3.18.0